+++ pcs/node.py
@@ -1,8 +1,5 @@
 import json
-from typing import (
-    Any,
-    Optional,
-)
+from typing import Any
 
 import pcs.lib.pacemaker.live as lib_pacemaker
 from pcs import utils
@@ -12,10 +9,10 @@ from pcs.cli.common.errors import (
     CmdLineInputError,
 )
 from pcs.cli.common.parse_args import (
+    OUTPUT_FORMAT_OPTION,
     Argv,
     InputModifiers,
     KeyValueParser,
-    ModifierValueType,
 )
 
 
@@ -23,20 +20,19 @@ def node_attribute_cmd(lib: Any, argv: Argv, modifiers: InputModifiers) -> None:
     """
     Options:
       * -f - CIB file (in lib wrapper)
-      * --force - allows not unique recipient values
-      * --name - specify attribute name to filter out
+      * --force - no error if attribute to delete doesn't exist
+      * --name - specify attribute name for filter
+      * --output-format - supported formats: text, cmd, json
     """
     del lib
-    modifiers.ensure_only_supported("-f", "--force", "--name")
-    if modifiers.get("--name") and len(argv) > 1:
+    modifiers.ensure_only_supported(
+        "-f", "--force", "--name", output_format_supported=True
+    )
+    if len(argv) < 2 or modifiers.is_specified_any(
+        ["--name", OUTPUT_FORMAT_OPTION]
+    ):
         raise CmdLineInputError()
-    if not argv:
-        attribute_show_cmd(filter_attr=modifiers.get("--name"))
-    elif len(argv) == 1:
-        attribute_show_cmd(argv.pop(0), filter_attr=modifiers.get("--name"))
-    else:
-        # --force is used only when setting attributes
-        attribute_set_cmd(argv.pop(0), argv)
+    attribute_set_cmd(argv.pop(0), argv)
 
 
 def node_utilization_cmd(
@@ -45,10 +41,15 @@ def node_utilization_cmd(
     """
     Options:
       * -f - CIB file (in lib wrapper)
-      * --name - specify attribute name to filter out
+      * --name - specify attribute name for filter
+      * --output-format - supported formats: text, cmd, json
     """
-    modifiers.ensure_only_supported("-f", "--name")
-    if modifiers.get("--name") and len(argv) > 1:
+    modifiers.ensure_only_supported(
+        "-f", "--name", output_format_supported=True
+    )
+    if len(argv) < 2 or modifiers.is_specified_any(
+        ["--name", OUTPUT_FORMAT_OPTION]
+    ):
         raise CmdLineInputError()
     utils.print_warning_if_utilization_attrs_has_no_effect(
         PropertyConfigurationFacade.from_properties_dtos(
@@ -56,12 +57,7 @@ def node_utilization_cmd(
             lib.cluster_property.get_properties_metadata(),
         )
     )
-    if not argv:
-        print_node_utilization(filter_name=modifiers.get("--name"))
-    elif len(argv) == 1:
-        print_node_utilization(argv.pop(0), filter_name=modifiers.get("--name"))
-    else:
-        set_node_utilization(argv.pop(0), argv)
+    set_node_utilization(argv.pop(0), argv)
 
 
 def node_maintenance_cmd(
@@ -154,49 +150,6 @@ def set_node_utilization(node: str, argv: Argv) -> None:
     utils.replace_cib_configuration(cib)
 
 
-def print_node_utilization(
-    filter_node: Optional[str] = None,
-    filter_name: ModifierValueType = None,
-) -> None:
-    """
-    Commandline options:
-      * -f - CIB file
-    """
-    cib = utils.get_cib_dom()
-
-    node_element_list = cib.getElementsByTagName("node")
-
-    if (
-        filter_node
-        and filter_node
-        not in [
-            node_element.getAttribute("uname")
-            for node_element in node_element_list
-        ]
-        and (
-            utils.usefile
-            or filter_node
-            not in [
-                node_attrs.name
-                for node_attrs in utils.getNodeAttributesFromPacemaker()
-            ]
-        )
-    ):
-        utils.err(f"Unable to find a node: {filter_node}")
-
-    utilization = {}
-    for node_el in node_element_list:
-        node = node_el.getAttribute("uname")
-        if filter_node is not None and node != filter_node:
-            continue
-        util_str = utils.get_utilization_str(node_el, filter_name)
-        if util_str:
-            utilization[node] = util_str
-    print("Node Utilization:")
-    for node in sorted(utilization):
-        print(f" {node}: {utilization[node]}")
-
-
 def node_pacemaker_status(
     lib: Any, argv: Argv, modifiers: InputModifiers
 ) -> None:
@@ -209,21 +162,6 @@ def node_pacemaker_status(
     print(json.dumps(lib_pacemaker.get_local_node_status(utils.cmd_runner())))
 
 
-def attribute_show_cmd(
-    filter_node: Optional[str] = None,
-    filter_attr: ModifierValueType = None,
-) -> None:
-    """
-    Commandline options:
-      * -f - CIB file (in lib wrapper)
-    """
-    node_attributes = utils.get_node_attributes(
-        filter_node=filter_node, filter_attr=filter_attr
-    )
-    print("Node Attributes:")
-    attribute_print(node_attributes)
-
-
 def attribute_set_cmd(node: str, argv: Argv) -> None:
     """
     Commandline options:
@@ -232,14 +170,3 @@ def attribute_set_cmd(node: str, argv: Argv) -> None:
     """
     for name, value in KeyValueParser(argv).get_unique().items():
         utils.set_node_attribute(name, value, node)
-
-
-def attribute_print(node_attributes):
-    """
-    Commandline options: no options
-    """
-    for node in sorted(node_attributes.keys()):
-        line_parts = [" " + node + ":"]
-        for name, value in sorted(node_attributes[node].items()):
-            line_parts.append(f"{name}={value}")
-        print(" ".join(line_parts))